requirements.txt
Deprecated==1.2.18
img2pdf==0.6.0
lxml==5.3.1
packaging==24.2
pikepdf==9.5.2
pillow==11.1.0
pillow_heif==0.22.0
wrapt==1.17.2
